Refactored resilience module
