--- src/resilience/fallback/_module.ts
@@ -1,1 +1,0 @@
-export * from "@/resilience/fallback/fallback.js";
