+++ src/resilience/implementations/fallback/_module.ts
@@ -1,0 +1,1 @@
+export * from "@/resilience/implementations/fallback/fallback.js";
