+++ src/resilience/implementations/fallback/fallback.test.ts
@@ -2,7 +2,7 @@ import { beforeEach, describe, expect, test, vi } from "vitest";
 
 import { type Use } from "@/middleware/contracts/_module.js";
 import { useFactory } from "@/middleware/implementations/_module.js";
-import { fallback } from "@/resilience/fallback/fallback.js";
+import { fallback } from "@/resilience/implementations/fallback/fallback.js";
 
 describe("function: fallback", () => {
     let use: Use;
