--- src/resilience/retry-interval/_module.ts
@@ -1,1 +1,0 @@
-export * from "@/resilience/retry-interval/retry-interval.js";
