--- src/resilience/timeout/_module.ts
@@ -1,1 +1,0 @@
-export * from "@/resilience/timeout/timeout.js";
